Move progressBar
