+++ src/components/pages/Providers/Providers.tsx
@@ -1,6 +1,6 @@
 import { Footer } from '#components/common/Footer'
 import { Header } from '#components/common/Header'
-import ProgressBar from '#components/common/Header/ProgressBar'
+import ProgressBar from '#components/common/ProgressBar/ProgressBar'
 import { FCC } from '#types/types'
 import { MDXProvider } from '@mdx-js/react'
 import React from 'react'
